+++ support/homebrew-cask/container-desktop.rb
@@ -1,11 +1,8 @@
 cask "container-desktop" do
-  arch arm: "arm64", intel: "x64"
-
   version "5.2.13"
-  sha256 arm:   "4e14f9b0b0f936aea1ca3433ba58fe963bbcaad331880586062e8d37c7b9df51",
-         intel: "43a4cbe20b06d439006a8c9c6cf826400be859f7b66fc42aa777466edef8d017"
+  sha256 "4e14f9b0b0f936aea1ca3433ba58fe963bbcaad331880586062e8d37c7b9df51"
 
-  url "https://github.com/iongion/container-desktop/releases/download/#{version}/container-desktop-#{arch}-#{version}.dmg",
+  url "https://github.com/iongion/container-desktop/releases/download/#{version}/container-desktop-arm64-#{version}.dmg",
       verified: "github.com/iongion/container-desktop/"
   name "Container Desktop"
   desc "General purpose container operations"
@@ -17,6 +14,7 @@ cask "container-desktop" do
   end
 
   auto_updates true
+  depends_on arch: :arm64
   depends_on macos: ">= :catalina"
 
   app "Container Desktop.app"